+++ src/main/java/org/avenir/reinforce/RewardComparisonLearner.java
@@ -88,7 +88,7 @@ public class RewardComparisonLearner extends MultiArmBanditLearner {
 	}
 	
 	@Override
-	public void setReward(String actionId, int reward) {
+	public void setReward(String actionId, double reward) {
 		rewardStats.get(actionId).add(reward);
 		rewarded = true;
 		findAction(actionId).reward(reward);
